backend/requirements.txt
Flask==3.1.0
flask-cors==6.0.2
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
psycopg2-binary==2.9.11
python-dotenv==1.0.1
gunicorn==22.0.0
